requirements-pinned.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.1
Automat==20.2.0
base58==2.1.1
bcrypt==3.2.0
binaryornot==0.4.4
bitarray==1.2.2
bitstring==3.1.9
cached-property==1.5.2
cbor2==5.4.2.post1
certifi==2021.10.8
cffi==1.15.0
cfxdb==22.2.2
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
constantly==15.1.0
cookiecutter==1.7.3
cryptography==36.0.1
cytoolz==0.11.2
docker==5.0.3
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.3
flatbuffers==2.0
frozenlist==1.3.0
h2==3.2.0
hashin==0.17.0
hexbytes==0.2.2
hkdf==0.0.3
hpack==3.0.0
humanize==4.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-resources==5.4.0
incremental==21.3.0
ipfshttpclient==0.8.0a2
iso8601==1.0.2
itsdangerous==2.1.0
Jinja2==3.0.3
jinja2-highlight==0.6.1
jinja2-time==0.2.0
jsonschema==3.2.0
lmdb==1.3.0
lru-dict==1.1.7
MarkupSafe==2.1.0
mistune==2.0.2
mnemonic==0.20
morphys==1.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.22.2
packaging==21.3
parsimonious==0.8.1
passlib==1.7.4
poyo==0.5.0
priority==1.3.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
PTable==0.9.2
py-cid @ git+https://github.com/crossbario/py-cid.git@e1dc52a43ced53845679405f78d92fb18f093653
py-ecc==5.2.0
py-eth-sig-utils==0.4.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash @ git+https://github.com/crossbario/py-multihash.git@86b54b8f9f0cf14c7370f38ae32f2b4b14f5b5e0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyGObject==3.42.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyQRCode==1.2.1
pyrsistent==0.18.1
python-baseconv==1.2.2
python-dateutil==2.8.2
python-slugify==6.1.0
python-snappy==0.6.1
PyTrie==0.4.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rlp==2.0.1
sdnotify==0.3.2
service-identity==21.1.0
setproctitle==1.2.2
setuptools==60.5.0
six==1.16.0
sortedcontainers==2.4.0
spake2==0.8
stringcase==1.2.0
tabulate==0.8.9
text-unidecode==1.3
toolz==0.11.2
treq==22.2.0
Twisted==22.1.0
txaio==22.2.1
txtorcon==21.1.0
typing_extensions==4.1.1
u-msgpack-python==2.7.1
ujson==5.1.0
urllib3==1.26.8
validate-email==1.3
varint==1.0.2
vmprof @ git+https://github.com/vmprof/vmprof-python.git@45a03f45b31a09dba99f77884f88318cc64db899
watchdog==2.1.6
wcwidth==0.2.5
web3==5.28.0
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.0.3
wsaccel==0.6.3
xbr==21.2.1
yapf==0.29.0
yarl==1.7.2
zipp==3.7.0
zlmdb==22.2.1
zope.interface==5.4.0
